few changes

+++ FrontEndAudToText/Models/WordResult.cs
@@ -2,7 +2,7 @@
 {
     public class WordResult
     {
-        public float start { get; set; }
+        public int start { get; set; }
         public float end { get; set; }
         public string text { get; set; }
     }
